// client-svelte/src/lib/generated/apiClient.ts
export class Download implements IDownload {
	downloadId?: string;
	torrentId?: string;
	torrent?: Torrent | undefined;
	path?: string;
	link?: string | undefined;
	added?: Date;
	downloadQueued?: Date | undefined;
	downloadStarted?: Date | undefined;
	downloadFinished?: Date | undefined;
	unpackingQueued?: Date | undefined;
	unpackingStarted?: Date | undefined;
	unpackingFinished?: Date | undefined;
	completed?: Date | undefined;
	retryCount?: number;
	error?: string | undefined;
	remoteId?: string | undefined;
	fileName?: string | undefined;
	bytesTotal?: number;
	bytesDone?: number;
	speed?: number;

	constructor(data?: IDownload) {
		if (data) {
			for (var property in data) {
				if (data.hasOwnProperty(property)) (<any>this)[property] = (<any>data)[property];
			}
		}
	}

	init(_data?: any) {
		if (_data) {
			this.downloadId = _data['downloadId'];
			this.torrentId = _data['torrentId'];
			this.torrent = _data['torrent'] ? Torrent.fromJS(_data['torrent']) : <any>undefined;
			this.path = _data['path'];
			this.link = _data['link'];
			this.added = _data['added'] ? new Date(_data['added'].toString()) : <any>undefined;
			this.downloadQueued = _data['downloadQueued']
				? new Date(_data['downloadQueued'].toString())
				: <any>undefined;
			this.downloadStarted = _data['downloadStarted']
				? new Date(_data['downloadStarted'].toString())
				: <any>undefined;
			this.downloadFinished = _data['downloadFinished']
				? new Date(_data['downloadFinished'].toString())
				: <any>undefined;
			this.unpackingQueued = _data['unpackingQueued']
				? new Date(_data['unpackingQueued'].toString())
				: <any>undefined;
			this.unpackingStarted = _data['unpackingStarted']
				? new Date(_data['unpackingStarted'].toString())
				: <any>undefined;
			this.unpackingFinished = _data['unpackingFinished']
				? new Date(_data['unpackingFinished'].toString())
				: <any>undefined;
			this.completed = _data['completed']
				? new Date(_data['completed'].toString())
				: <any>undefined;
			this.retryCount = _data['retryCount'];
			this.error = _data['error'];
			this.remoteId = _data['remoteId'];
			this.fileName = _data['fileName'];
			this.bytesTotal = _data['bytesTotal'];
			this.bytesDone = _data['bytesDone'];
			this.speed = _data['speed'];
		}
	}

	static fromJS(data: any): Download {
		data = typeof data === 'object' ? data : {};
		let result = new Download();
		result.init(data);
		return result;
	}

	toJSON(data?: any) {
		data = typeof data === 'object' ? data : {};
		data['downloadId'] = this.downloadId;
		data['torrentId'] = this.torrentId;
		data['torrent'] = this.torrent ? this.torrent.toJSON() : <any>undefined;
		data['path'] = this.path;
		data['link'] = this.link;
		data['added'] = this.added ? this.added.toISOString() : <any>undefined;
		data['downloadQueued'] = this.downloadQueued
			? this.downloadQueued.toISOString()
			: <any>undefined;
		data['downloadStarted'] = this.downloadStarted
			? this.downloadStarted.toISOString()
			: <any>undefined;
		data['downloadFinished'] = this.downloadFinished
			? this.downloadFinished.toISOString()
			: <any>undefined;
		data['unpackingQueued'] = this.unpackingQueued
			? this.unpackingQueued.toISOString()
			: <any>undefined;
		data['unpackingStarted'] = this.unpackingStarted
			? this.unpackingStarted.toISOString()
			: <any>undefined;
		data['unpackingFinished'] = this.unpackingFinished
			? this.unpackingFinished.toISOString()
			: <any>undefined;
		data['completed'] = this.completed ? this.completed.toISOString() : <any>undefined;
		data['retryCount'] = this.retryCount;
		data['error'] = this.error;
		data['remoteId'] = this.remoteId;
		data['fileName'] = this.fileName;
		data['bytesTotal'] = this.bytesTotal;
		data['bytesDone'] = this.bytesDone;
		data['speed'] = this.speed;
		return data;
	}
}

export interface IDownload {
	downloadId?: string;
	torrentId?: string;
	torrent?: Torrent | undefined;
	path?: string;
	link?: string | undefined;
	added?: Date;
	downloadQueued?: Date | undefined;
	downloadStarted?: Date | undefined;
	downloadFinished?: Date | undefined;
	unpackingQueued?: Date | undefined;
	unpackingStarted?: Date | undefined;
	unpackingFinished?: Date | undefined;
	completed?: Date | undefined;
	retryCount?: number;
	error?: string | undefined;
	remoteId?: string | undefined;
	fileName?: string | undefined;
	bytesTotal?: number;
	bytesDone?: number;
	speed?: number;
}